🆕 chore(nbt_1_5_1.xml): add nbt-1.5.1 library 🆕 chore(Minecraft_Server_Launcher_2.iml): add nbt-1.5.1 library to project dependencies 🆕 chore(nbt-1.5.1.jar): add nbt-1.5.1 library jar file 🐛 fix(AddWorldsPanel.class, AlertType.class, ButtonPanel.class): binary files differ This commit adds the nbt-1.5.1 library to the project dependencies. The library is added to the libraryTable in the .idea/libraries/nbt_1_5_1.xml file, and to the orderEntry in the Minecraft_Server_Launcher_2.iml file. The nbt-1.5.1.jar file is also added to the project. There are no changes to the source code, but the binary files
🔨 chore: update compiled class files
The compiled class files have been updated, but no changes have been made to the source code. This commit is simply to update the compiled files.

🐛 fix(AddWorldsPanel.java): add null check for icon.png file
✨ feat(AddWorldsPanel.java): add support for NBT library
The null check for the icon.png file prevents the application from crashing when the file is missing. The support for the NBT library allows the application to read and write NBT files, which is useful for Minecraft worlds. Additionally, issue #8 is fixed by adding a boolean to check the content's type, and issue #22 and #24 are fixed by adding checks for the existence of the icon.png file.

🐛 fix(AddWorldsPanel.java): add Nbt object initialization
The Nbt object was not initialized before being used, which caused a NullPointerException. This commit initializes the Nbt object to fix the issue.

🐛 fix(WorldCopyHandler.java): fix issue with directory write
The directory write was fixed by changing the absolute path of the extractedDirTemp variable to go one folder deeper. The previous implementation was writing to the wrong directory.