🔥 chore(gui): remove binary files The commit removes binary files that are not needed in the repository.
🔨 chore: update compiled class files
This commit updates the compiled class files for the Minecraft Server Launcher 2 project. There are no changes to the source code, only the compiled binaries have been updated.

🚀 feat(AddWorldsPanel.java): add server selection to WorldCopyHandler constructor
This commit adds the server selection to the WorldCopyHandler constructor in three different places in the AddWorldsPanel.java file. This allows the WorldCopyHandler to know which server to copy the world to.

🐛 fix(AddWorldsPanel.java): pass selected server index to WorldCopyHandler constructor
The WorldCopyHandler constructor was not being passed the selected server index, which caused the application to crash when attempting to copy a world. The fix involves passing the selected server index to the constructor, which allows the WorldCopyHandler to properly handle the copy operation.

🐛 fix(ConfigStuffPanel.java): remove unused comboBoxSelectedIndex variable
🐛 fix(ServerSelector.java): fix startCopying and cancel buttons to use the correct file path
✨ feat(ServerSelector.java): add selectedIndex variable to store the selected index of the server
The comboBoxSelectedIndex variable is no longer used and has been removed. The startCopying and cancel buttons in ServerSelector.java have been fixed to use the correct file path. The selectedIndex variable has been added to store the selected index of the server.

🔒 refactor(ServerProperties.java): use Config object to get server properties file path
This commit refactors the ServerProperties class to use a Config object to get the path to the server properties file instead of using a static method from the ConfigStuffPanel class. This improves the code's maintainability and reduces coupling between classes.

🔨 refactor(WorldCopyHandler.java): add constructor parameter for config index
This commit adds a constructor parameter for the config index to the WorldCopyHandler class. This allows the class to be more flexible and work with different configurations. The config index is used to retrieve the path to the server folder from the configuration file.

🐛 fix(WorldCopyHandler.java): replace ConfigStuffPanel.getServPath() with config.getData().get(configIndex).getPathToServerFolder()
The ConfigStuffPanel.getServPath() method has been replaced with config.getData().get(configIndex).getPathToServerFolder() to ensure that the correct path to the server folder is used. This change ensures that the correct directory is checked for the existence of the server world directory and that the correct directory is used when copying the world directory to the server world directory.